requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aiohttp==3.9.5
flask==3.0.3
psycopg2-binary==2.9.9